Refactor 3 Move all interaction and execution related functions into action
